atol en malloc y free arreglado

+++ c/so/shell/src/malloc.c
@@ -16,7 +16,7 @@ int cmd_malloc(char **arg)
 		return 0;
 	}
 	
-	int size = atoi(arg[1]);
+	int size = atol(arg[1]);
 	time_t t = time(0);
 	void *p = malloc(size);	
 	if(p==NULL) {
